fix: Fix more useless comparisons and range checks (Refs PMAT-802)
- Fix useless >= 0 comparisons for unsigned types in:
  - cuda/executor/core.rs (pool_hits)
  - cuda/executor/layers/graphed.rs (decode_token_count)
  - cuda/executor/tests.rs (num_devices)
  - gguf/batch_scheduler.rs (avg_wait_time)
- Replace i8 range checks (always true) with simple assertions in:
  - quantize/tests/part_12.rs
  - quantize/tests/part_15.rs
  - quantize/tests/part_17.rs
  - quantize/tests/part_21.rs
  - tests/quantize_property_tests.rs
  - tests/proptest_q4_simd.rs
  - tests/cuda_deep_coverage.rs

+++ src/quantize/tests/part_05.rs
@@ -2377,7 +2377,7 @@ fn test_q4_0_fused_vs_dequantize_matmul() {
     for i in 0..16 {
         // Put increasing value in low nibble (0,1,2,...,15)
         // Put 8 in high nibble so high_quant = 8 - 8 = 0
-        data[2 + i] = ((8 << 4) | (i as u8)) as u8;
+        data[2 + i] = ((8 << 4) | (i as u8));
     }
 
     // Row 1: scale=2.0, all quants = 0x99 (low=9, high=9)
